enhancement: user menu dividers
